rename token module
