+++ remme/__main__.py
@@ -20,6 +20,7 @@ from remme.atomic_swap_tp.handler import AtomicSwapHandler
 from remme.certificate.handler import CertificateHandler
 from remme.account.handler import AccountHandler
 
+
 from remme.shared.logging import setup_logging
