+++ remme/account/client.py
@@ -15,20 +15,20 @@
 
 import json
 
-from remme.protos.account_pb2 import AccountMethod, GenesisPayload, TransferPayload, Account
+from remme.protos.token_pb2 import TokenMethod, GenesisPayload, TransferPayload
 from remme.shared.basic_client import BasicClient
+from remme.account.handler import AcountHandler
 from remme.shared.exceptions import KeyNotFound
 
-
-from .handler import AccountHandler
+from remme.protos.token_pb2 import Account
 
 
 class AccountClient(BasicClient):
     def __init__(self):
-        super().__init__(AccountHandler)
+        super().__init__(AcountHandler)
 
     def _send_transaction(self, method, data_pb, extra_addresses_input_output):
-        addresses_input_output = [self.make_address_from_data(self._signer.get_public_key().as_hex())]
+        addresses_input_output = [self.get_user_address()]
         if extra_addresses_input_output:
             addresses_input_output += extra_addresses_input_output
         return super()._send_transaction(method, data_pb, addresses_input_output)
@@ -59,9 +59,7 @@ class AccountClient(BasicClient):
         extra_addresses_input_output = [address_to]
         transfer = self.get_transfer_payload(address_to, value)
 
-        status = self._send_transaction(AccountMethod.TRANSFER, transfer, extra_addresses_input_output)
-
-        return json.loads(status)
+        return self._send_transaction(TokenMethod.TRANSFER, transfer, extra_addresses_input_output)
 
     def get_account(self, address):
         account = Account()
