+++ remme/account/handler.py
@@ -16,48 +16,46 @@
 import logging
 from sawtooth_sdk.processor.exceptions import InvalidTransaction
 
-from remme.protos.account_pb2 import Account, GenesisStatus, AccountMethod, GenesisPayload, \
+from remme.protos.token_pb2 import Account, GenesisStatus, TokenMethod, GenesisPayload, \
     TransferPayload
+from remme.settings import SETTINGS_KEY_GENESIS_OWNERS, GENESIS_ADDRESS, ZERO_ADDRESS
+from remme.settings.helper import _get_setting_value
 from remme.shared.basic_handler import *
 from remme.shared.singleton import singleton
 
 LOGGER = logging.getLogger(__name__)
 
-ZERO_ADDRESS = '0' * 64
-
 FAMILY_NAME = 'account'
 FAMILY_VERSIONS = ['0.1']
 
+
+def get_account_by_address(context, address):
+    account = get_data(context, Account, address)
+    if account is None:
+        return Account()
+    return account
 
 # TODO: ensure receiver_account.balance += transfer_payload.amount is within uint64
 @singleton
 class AccountHandler(BasicHandler):
     def __init__(self):
         super().__init__(FAMILY_NAME, FAMILY_VERSIONS)
-        self.zero_address = self.make_address(ZERO_ADDRESS)
 
     def get_state_processor(self):
         return {
-            AccountMethod.TRANSFER: {
+            TokenMethod.TRANSFER: {
                 'pb_class': TransferPayload,
                 'processor': self._transfer
             },
-            AccountMethod.GENESIS: {
+            TokenMethod.GENESIS: {
                 'pb_class': GenesisPayload,
                 'processor': self._genesis
             }
         }
 
-    def get_account_by_pub_key(self, context, pub_key):
-        address = self.make_address_from_data(pub_key)
-        account = get_data(context, Account, address)
-        if account is None:
-            return address, Account()
-        return address, account
-
     def _genesis(self, context, pub_key, genesis_payload):
-        signer_key, account = self.get_account_by_pub_key(context, pub_key)
-        genesis_status = get_data(context, GenesisStatus, self.zero_address)
+        signer_key = self.make_address_from_data(pub_key)
+        genesis_status = get_data(context, GenesisStatus, GENESIS_ADDRESS)
         if not genesis_status:
             genesis_status = GenesisStatus()
         elif genesis_status.status:
@@ -69,15 +67,30 @@ class AccountHandler(BasicHandler):
                     .format(genesis_payload.total_supply, signer_key))
         return {
             signer_key: account,
-            self.zero_address: genesis_status
+            GENESIS_ADDRESS: genesis_status
         }
 
     def _transfer(self, context, pub_key, transfer_payload):
-        signer_key, signer_account = self.get_account_by_pub_key(context, pub_key)
-        if self.zero_address in [transfer_payload.address_to, signer_key]:
-            raise InvalidTransaction("Zero address cannot involve in any operation.")
+        address = self.make_address_from_data(pub_key)
+        return self._transfer_from_address(context, address, transfer_payload)
+
+    def _transfer_from_address(self, context, address, transfer_payload):
+        signer_key = address
+        signer_account = get_account_by_address(context, address)
+
+        if not transfer_payload.address_to.startswith(self._prefix) \
+                and transfer_payload.address_to not in [ZERO_ADDRESS]:
+            raise InvalidTransaction("Receiver address has to be of an account type")
+
         if signer_key == transfer_payload.address_to:
             raise InvalidTransaction("Account cannot send tokens to itself.")
+
+        # TODO transfer from genesis address using SETTINGS_KEY_GENESIS_OWNERS list of allowed addresses(0x0)
+        # genesis_members_str = _get_setting_value(context, SETTINGS_KEY_GENESIS_OWNERS)
+        # if not genesis_members_str:
+        #     raise InvalidTransaction('REMchain is not configured to process genesis transfers.')
+        #
+        # genesis_members_list = genesis_members_str.split()
 
         receiver_account = get_data(context, Account, transfer_payload.address_to)
 
@@ -87,7 +100,7 @@ class AccountHandler(BasicHandler):
             signer_account = Account()
 
         if signer_account.balance < transfer_payload.value:
-            raise InvalidTransaction("Not enough transferable balance. Signer's current balance: {}"
+            raise InvalidTransaction("Not enough transferable balance. Sender's current balance: {}"
                                      .format(signer_account.balance))
 
         receiver_account.balance += transfer_payload.value
