+++ remme/rest_api/token.py
@@ -17,6 +17,7 @@ import re
 from remme.account.client import AccountClient
 
 
+
 def get(pub_key_user):
     client = AccountClient()
     address = client.make_address_from_data(pub_key_user)
