--- remme/token_tp/handler.py
@@ -1,116 +1,0 @@
-# Copyright 2018 REMME
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ------------------------------------------------------------------------
-
-import logging
-from sawtooth_sdk.processor.exceptions import InvalidTransaction
-
-from remme.protos.token_pb2 import Account, GenesisStatus, TokenMethod, GenesisPayload, \
-    TransferPayload
-from remme.settings import SETTINGS_KEY_GENESIS_OWNERS, GENESIS_ADDRESS, ZERO_ADDRESS
-from remme.settings.helper import _get_setting_value
-from remme.shared.basic_handler import *
-from remme.shared.singleton import singleton
-
-LOGGER = logging.getLogger(__name__)
-
-FAMILY_NAME = 'token'
-FAMILY_VERSIONS = ['0.1']
-
-
-def get_account_by_address(context, address):
-    account = get_data(context, Account, address)
-    if account is None:
-        return Account()
-    return account
-
-# TODO: ensure receiver_account.balance += transfer_payload.amount is within uint64
-@singleton
-class TokenHandler(BasicHandler):
-    def __init__(self):
-        super().__init__(FAMILY_NAME, FAMILY_VERSIONS)
-
-    def get_state_processor(self):
-        return {
-            TokenMethod.TRANSFER: {
-                'pb_class': TransferPayload,
-                'processor': self._transfer
-            },
-            TokenMethod.GENESIS: {
-                'pb_class': GenesisPayload,
-                'processor': self._genesis
-            }
-        }
-
-    def _genesis(self, context, pub_key, genesis_payload):
-        signer_key = self.make_address_from_data(pub_key)
-        genesis_status = get_data(context, GenesisStatus, GENESIS_ADDRESS)
-        if not genesis_status:
-            genesis_status = GenesisStatus()
-        elif genesis_status.status:
-            raise InvalidTransaction('Genesis is already initialized.')
-        genesis_status.status = True
-        account = Account()
-        account.balance = genesis_payload.total_supply
-        LOGGER.info('Generated genesis transaction. Issued {} tokens to address {}'
-                    .format(genesis_payload.total_supply, signer_key))
-        return {
-            signer_key: account,
-            GENESIS_ADDRESS: genesis_status
-        }
-
-    def _transfer(self, context, pub_key, transfer_payload):
-        address = self.make_address_from_data(pub_key)
-        return self._transfer_from_address(context, address, transfer_payload)
-
-    def _transfer_from_address(self, context, address, transfer_payload):
-        signer_key = address
-        signer_account = get_account_by_address(context, address)
-
-        if not transfer_payload.address_to.startswith(self._prefix) \
-                and transfer_payload.address_to not in [ZERO_ADDRESS]:
-            raise InvalidTransaction("Receiver address has to be of an account type")
-
-        if signer_key == transfer_payload.address_to:
-            raise InvalidTransaction("Account cannot send tokens to itself.")
-
-        # TODO transfer from genesis address using SETTINGS_KEY_GENESIS_OWNERS list of allowed addresses(0x0)
-        # genesis_members_str = _get_setting_value(context, SETTINGS_KEY_GENESIS_OWNERS)
-        # if not genesis_members_str:
-        #     raise InvalidTransaction('REMchain is not configured to process genesis transfers.')
-        #
-        # genesis_members_list = genesis_members_str.split()
-
-        receiver_account = get_data(context, Account, transfer_payload.address_to)
-
-        if not receiver_account:
-            receiver_account = Account()
-        if not signer_account:
-            signer_account = Account()
-
-        if signer_account.balance < transfer_payload.value:
-            raise InvalidTransaction("Not enough transferable balance. Sender's current balance: {}"
-                                     .format(signer_account.balance))
-
-        receiver_account.balance += transfer_payload.value
-        signer_account.balance -= transfer_payload.value
-
-        LOGGER.info('Transferred {} tokens from {} to {}'.format(transfer_payload.value,
-                                                                 signer_key,
-                                                                 transfer_payload.address_to))
-
-        return {
-            signer_key: signer_account,
-            transfer_payload.address_to: receiver_account
-        }
